Combine teamStats and pitcherData updating
